backend/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
